+++ apis/v1beta1/zz_generated.pculist.go
@@ -29,11 +29,3 @@ func (p *ClusterProviderConfigUsageList) GetItems() []resource.ProviderConfigUsa
 	return items
 }
 
-// GetItems of this ProviderConfigUsageList.
-func (p *ProviderConfigUsageList) GetItems() []resource.ProviderConfigUsage {
-	items := make([]resource.ProviderConfigUsage, len(p.Items))
-	for i := range p.Items {
-		items[i] = &p.Items[i]
-	}
-	return items
-}